+++ poteto.go
@@ -56,6 +56,14 @@ func (p *poteto) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	// get from cache & reset context
 	ctx := p.initializeContext(w, r)
 
+	// get and SetRequestId
+	// slow -> should be middleware?
+	reqId := ctx.RequestId()
+	ctx.Set(constant.STORE_REQUEST_ID, reqId)
+	if id := ctx.GetRequest().Header.Get(constant.HEADER_X_REQUEST_ID); id == "" {
+		ctx.GetResponse().Header().Set(constant.HEADER_X_REQUEST_ID, reqId)
+	}
+
 	routes := p.router.GetRoutesByMethod(r.Method)
 
 	targetRoute, httpParam := routes.Search(r.URL.Path)